pyproject.toml
[project]
name = "smart-expense-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.76.0",
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.3.1",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
